Convert id to int

+++ app/src/main/java/t3waii/tasklists/CreatedTasksFragment.java
@@ -134,8 +134,8 @@ public class CreatedTasksFragment extends TasksFragment implements PopupMenu.OnM
 
     @Override
     public boolean onMenuItemClick(MenuItem item) {
-        Long taskId = item.getIntent().getLongExtra("taskId", 0);
-        Log.d(TAG, "taskId:" + Long.toString(taskId)); //getActionView().toString()
+        int taskId = item.getIntent().getIntExtra("taskId", 0);
+        Log.d(TAG, "taskId:" + Integer.toString(taskId)); //getActionView().toString()
 
         switch (item.getItemId()) {
             case R.id.edit_task:
